starlark::analysis

Trait AstModuleLint

pub trait AstModuleLint {
    // Required method
    fn lint(&self, globals: Option<&HashSet<String>>) -> Vec<Lint>;
}

Run the linter.

Required Methods

fn lint(&self, globals: Option<&HashSet<String>>) -> Vec<Lint>

Run a static linter over the module. If the complete set of global variables are known they can be passed as the globals argument, resulting in name-resolution lint errors. The precise checks run by the linter are not considered stable between versions.