starlark

Module any

Methods that build upon the Any trait.

Traits

AnyLifetime

Like Any, but while Any requires 'static, this version allows a lifetime parameter.

ProvidesStaticType

Provides access to the same type as Self but with all lifetimes dropped to 'static (including lifetimes of parameters).

Derive Macros

ProvidesStaticType

Derive the ProvidesStaticType trait. Requires the type has no type arguments, no constant arguments, and at most one lifetime argument.