starlark::any

Trait AnyLifetime

pub trait AnyLifetime<'a>: ProvidesStaticTypeSealed<'a> + 'a {
    // Required methods
    fn static_type_id() -> TypeId
       where Self: Sized;
    fn static_type_of(&self) -> TypeId;
}

Like Any, but while Any requires 'static, this version allows a lifetime parameter.

Code using this trait is unsafe if your implementation of the inner methods do not meet the invariants listed. Therefore, it is recommended you use one of the helper macros.

You cannot implement this trait directly. You should instead implement ProvidesStaticType, usually via the derive macro:

use starlark::any::ProvidesStaticType;
#[derive(ProvidesStaticType)]
struct Foo1();
#[derive(ProvidesStaticType)]
struct Foo2<'a>(&'a ());

If your data type is not of the form Foo or Foo<'v> you may need to implement ProvidesStaticType directly.

use starlark::any::ProvidesStaticType;
struct Baz<T: Display>(T);
unsafe impl<'a, T> ProvidesStaticType<'a> for Baz<T>
where
    T: ProvidesStaticType<'a> + Display,
    T::StaticType: Display + Sized,
{
    type StaticType = Baz<T::StaticType>;
}

Required Methods

fn static_type_id() -> TypeId
where Self: Sized,

Must return the TypeId of Self but where the lifetimes are changed to 'static. Must be consistent with static_type_of.

fn static_type_of(&self) -> TypeId

Must return the TypeId of Self but where the lifetimes are changed to 'static. Must be consistent with static_type_id. Must not consult the self parameter in any way.

Implementations

impl<'a> dyn AnyLifetime<'a>

pub fn is<T: AnyLifetime<'a>>(&self) -> bool

Is the value of type T.

pub fn downcast_ref<T: AnyLifetime<'a>>(&self) -> Option<&T>

Downcast a reference to type T, or return None if it is not the right type.

pub fn downcast_mut<T: AnyLifetime<'a>>(&mut self) -> Option<&mut T>

Downcast a mutable reference to type T, or return None if it is not the right type.