starlark::any

Trait ProvidesStaticType

pub unsafe trait ProvidesStaticType<'a> {
    type StaticType: 'static + ?Sized;
}

Provides access to the same type as Self but with all lifetimes dropped to 'static (including lifetimes of parameters).

This type is usually implemented with #[derive(ProvidesStaticType)].

Required Associated Types

type StaticType: 'static + ?Sized

Same type as Self but with lifetimes dropped to 'static.

The trait is unsafe because if this is implemented incorrectly, the program might not work correctly.