starlark

Module bazel

This file provides an implementation of LspContext specifically aimed at the use in a Bazel project. You can invoke it by using starlark --lsp --bazel. Note that only --lsp mode is supported.

This module is temporary, for the purpose of rapid iteration while the LSP interface develops. After the API of the LspContext trait stabilizes, this module will be removed, and extracted to its own project.

Modules

label 🔒

Module for parsing bazel labels

Structs

BazelContext 🔒

FilesystemCompletionOptions 🔒

Options for resolving filesystem completions.

Enums

ContextError 🔒

FilesystemCompletionRoot 🔒

Starting point for resolving filesystem completions.

FilesystemFileCompletionOptions 🔒

RenderLoadError 🔒

Errors when LspContext::render_as_load() cannot render a given path.

ResolveLoadError 🔒

Errors when LspContext::resolve_load() cannot resolve a given path.

Functions

main 🔒