starlark::coerce

Trait Coerce

pub unsafe trait Coerce<To: ?Sized> { }

A marker trait such that the existence of From: Coerce<To> implies that From can be treat as To without any data manipulation. Particularly useful for containers, e.g. Vec<From> can be treated as Vec<To> in O(1). If such an instance is available, you can use coerce to perform the conversion.

Importantly, you must make sure Rust does not change the type representation between the different types (typically using a repr directive), and it must be safe for the From to be treated as To, namely same (or less restrictive) alignment, no additional invariants, value can be dropped as To.

If you only need coerce on newtype references, then the ref-cast crate provides that, along with automatic derivations (no unsafe required).