starlark

Module collections

Defines SmallMap and SmallSet - collections with deterministic iteration and small memory footprint.

These structures use vector backed storage if there are only a few elements, and and index for larger collections. The API mirrors standard Rust collections.

Structs

Hashed

A key and its hash.

IntoIter

Iterator that moves entries out of a SmallMap.

Iter

Iterator over a small map entry references.

IterMut

Iterator over a small map mutable entry references.

SmallMap

A map with deterministic iteration order.

SmallSet

An memory-efficient set with deterministic order, based on SmallMap.

StarlarkHashValue

A hash value.

StarlarkHasher

A hasher used by Starlark implementation.

Traits

Equivalent

Key equivalence trait.