starlark::collections

Struct SmallMap

#[repr(C)]
pub struct SmallMap<K, V> { /* private fields */ }

A map with deterministic iteration order.

This map is similar to indexmap::IndexMap with the following differences:

Small hashes are stored next to keys

Index is not created for small maps

Implementations

impl<K, V> SmallMap<K, V>

pub const fn new() -> SmallMap<K, V>

Empty map.

pub fn with_capacity(n: usize) -> SmallMap<K, V>

Create an empty map with specified capacity.

pub fn maybe_drop_index(&mut self)

Drop the index if the map is too small, and the index is not really needed.

We don’t allocate index prematurely when we add entries the map, but we keep it allocated when we remove entries from the map.

This function allows to reclaim memory after some entries are removed.

pub fn keys(&self) -> Keys<'_, K, V>

Key references iterator.

pub fn values(&self) -> Values<'_, K, V>

Value references iterator.

pub fn into_keys(self) -> IntoKeys<K, V>

Key owned iterator.

pub fn into_values(self) -> IntoValues<K, V>

Value owned iterator.

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V>

Mutable value references iterator.

pub fn iter(&self) -> Iter<'_, K, V>

Entry references iterator.

pub fn iter_hashed(&self) -> IterHashed<'_, K, V>

Entry references with hashes iterator.

pub fn into_iter_hashed(self) -> IntoIterHashed<K, V>

Entries with hashes iterator.

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Mutable entry references iterator.

pub fn iter_mut_unchecked(&mut self) -> IterMutUnchecked<'_, K, V>

Mutable entry references iterator, with mutable key references.

This operation is memory safe, but otherwise no guarantees if keys are mutated inconsistently (hash or equality changes).

pub fn get_hashed<Q>(&self, key: Hashed<&Q>) -> Option<&V>
where Q: Equivalent<K> + ?Sized,

Query the map by a prehashed key.

pub fn get_hashed_by_value<Q>(&self, key: Hashed<Q>) -> Option<&V>
where Q: Equivalent<K>,

Same as get_hashed, byt takes key by value instead of by reference. Sometimes it generates slightly better code for small values.

pub fn get<Q>(&self, key: &Q) -> Option<&V>
where Q: Hash + Equivalent<K> + ?Sized,

Query the map by a given key.

pub fn get_full<Q>(&self, key: &Q) -> Option<(usize, &K, &V)>
where Q: Hash + Equivalent<K> + ?Sized,

Query the map by a given key, return an index of the entry along with the entry key and value.

pub fn get_full_hashed<Q>(&self, key: Hashed<&Q>) -> Option<(usize, &K, &V)>
where Q: Equivalent<K> + ?Sized,

Query the map by a given key, return an index of the entry along with the entry key and value.

pub fn get_index_of_hashed<Q>(&self, key: Hashed<&Q>) -> Option<usize>
where Q: Equivalent<K> + ?Sized,

Find the index of the given hashed key.

pub fn get_index_of_hashed_by_value<Q>(&self, key: Hashed<Q>) -> Option<usize>
where Q: Equivalent<K>,

Get the index of the entry given a hashed key.

pub fn get_index(&self, index: usize) -> Option<(&K, &V)>

Find an entry by an index.

pub fn get_index_of<Q>(&self, key: &Q) -> Option<usize>
where Q: Hash + Equivalent<K> + ?Sized,

The an entry index by a given key.

pub fn get_mut_hashed<Q>(&mut self, key: Hashed<&Q>) -> Option<&mut V>
where Q: Equivalent<K> + ?Sized,

Find a mutable value by a hashed key.

pub fn get_mut<Q>(&mut self, key: &Q) -> Option<&mut V>
where Q: Hash + Equivalent<K> + ?Sized,

Find the entry by a given key.

pub fn contains_key_hashed<Q>(&self, key: Hashed<&Q>) -> bool
where Q: Equivalent<K> + ?Sized,

Find if an entry by a given prehashed key exists.

pub fn contains_key_hashed_by_value<Q>(&self, key: Hashed<Q>) -> bool
where Q: Equivalent<K>,

Find if an entry by a given hashed key exists.

pub fn contains_key<Q>(&self, key: &Q) -> bool
where Q: Hash + Equivalent<K> + ?Sized,

Find if an entry by a given key exists.

pub fn reserve(&mut self, additional: usize)

Reserve capacity for at least additional more elements to be inserted.

pub fn capacity(&self) -> usize

Current map capacity.

pub fn first(&self) -> Option<(&K, &V)>

Returns a reference to the first key-value pair.

pub fn last(&self) -> Option<(&K, &V)>

Returns a reference to the last key-value pair.

pub fn insert_hashed_unique_unchecked( &mut self, key: Hashed<K>, val: V, ) -> (&K, &mut V)

Insert an entry into the map without checking for a duplicate key.

pub fn insert_hashed(&mut self, key: Hashed<K>, val: V) -> Option<V>
where K: Eq,

Insert a key-value pair into the map.

pub fn insert(&mut self, key: K, val: V) -> Option<V>
where K: Hash + Eq,

Insert a key-value pair into the map.

pub fn insert_unique_unchecked(&mut self, key: K, val: V) -> (&K, &mut V)
where K: Hash,

Insert a key-value pair into the map without checking for a duplicate key.

pub fn shift_remove_hashed<Q>(&mut self, key: Hashed<&Q>) -> Option<V>
where Q: Equivalent<K> + ?Sized,

Remove the entry for the key.

Time complexity of this operation is O(N) where N is the number of entries in the map.

pub fn shift_remove_hashed_entry<Q>( &mut self, key: Hashed<&Q>, ) -> Option<(K, V)>
where Q: Equivalent<K> + ?Sized,

Remove the entry for the key.

Time complexity of this operation is O(N) where N is the number of entries in the map.

pub fn shift_remove_index_hashed(&mut self, i: usize) -> Option<(Hashed<K>, V)>

Remove the entry by index. This is O(N) operation.

pub fn shift_remove_index(&mut self, i: usize) -> Option<(K, V)>

Remove the entry by index. This is O(N) operation.

pub fn shift_remove<Q>(&mut self, key: &Q) -> Option<V>
where Q: Hash + Equivalent<K> + ?Sized,

Remove the entry for the key.

Time complexity of this operation is O(N) where N is the number of entries in the map.

pub fn shift_remove_entry<Q>(&mut self, key: &Q) -> Option<(K, V)>
where Q: Hash + Equivalent<K> + ?Sized,

Remove the entry for the key.

Time complexity of this operation is O(N) where N is the number of entries in the map.

pub fn entry_hashed(&mut self, key: Hashed<K>) -> Entry<'_, K, V>
where K: Eq,

Get the entry (occupied or not) for the key.

pub fn pop(&mut self) -> Option<(K, V)>

Remove the last element.

pub fn entry(&mut self, key: K) -> Entry<'_, K, V>
where K: Eq + Hash,

Get the entry (occupied or not) for the key.

pub fn is_empty(&self) -> bool

Is the map empty?

pub fn len(&self) -> usize

Get the number of elements in the map.

pub fn clear(&mut self)

Remove all elements from the map.

Retain the capacity.

pub fn sort_keys(&mut self)
where K: Ord,

Sort entries by key.

pub fn eq_ordered(&self, other: &SmallMap<K, V>) -> bool
where K: PartialEq, V: PartialEq,

Equal if the keys and values are equal in the iteration order.

pub fn hash_ordered<H>(&self, state: &mut H)
where H: Hasher, K: Hash, V: Hash,

Hash entries in the iteration order.

Note, keys are not hashed, but previously computed hashes are hashed instead.

pub fn reverse(&mut self)

Reverse the iteration order of the map.

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&K, &mut V) -> bool,

Retains only the elements specified by the predicate.