starlark

Module debug

Provides debug-related functionality and utilities.

Structs

EvaluateExprInfo

Information about expression evaluation result

InspectVariableInfo

Information about variable child “sub-values”

ResolvedBreakpoints

Breakpoints resolved to their spans.

ScopesInfo

Information about the variables scopes

Variable

Information about a “structural variable” inspected by a debugger this currently has DAP-like semantic meaning that every complex object returned by debugger from the stack or from the heap can be broken down into “variables” this is how structured data is managed by the debugger. Something similar to LLDB’s SBValue

VariablePath

Represents a variable’s “access path” for a local variable or watch expression.

VariablesInfo

Information about variables in scope.

Enums

PathSegment

Represents a segment in an access expression.

Scope

Represents the scope of a variable.

StepKind

The kind of debugger step, used for next/stepin/stepout requests.

Traits

DapAdapter

The DapAdapter accepts DAP requests and updates the hooks in the running evaluator.

DapAdapterClient

The DapAdapterClient is implemented by the user and provides functionality required by the DapAdapter.

DapAdapterEvalHook

This is sort of the evaluation side of the DapAdapter. It’s expected that these are on different threads (the starlark evaluation is single-threaded, so certainly the DapAdapter itself doesn’t do interesting things there).

Functions

dap_capabilities

The DAP capabilities that the adapter supports.

prepare_dap_adapter

Creates a DapAdapter and corresponding DapAdapterEvalHook.

resolve_breakpoints

Resolves the breakpoints to their FileSpan if possible.