starlark::debug

Trait DapAdapter

pub trait DapAdapter:
    Debug
    + Send
    + 'static {
    // Required methods
    fn set_breakpoints(
        &self,
        source: &str,
        breakpoints: &ResolvedBreakpoints,
    ) -> Result<()>;
    fn top_frame(&self) -> Result<Option<StackFrame>>;
    fn stack_trace(
        &self,
        args: StackTraceArguments,
    ) -> Result<StackTraceResponseBody>;
    fn scopes(&self) -> Result<ScopesInfo>;
    fn variables(&self) -> Result<VariablesInfo>;
    fn inspect_variable(
        &self,
        path: VariablePath,
    ) -> Result<InspectVariableInfo>;
    fn continue_(&self) -> Result<()>;
    fn step(&self, kind: StepKind) -> Result<()>;
    fn evaluate(&self, expr: &str) -> Result<EvaluateExprInfo>;
}

The DapAdapter accepts DAP requests and updates the hooks in the running evaluator.

Required Methods

fn set_breakpoints( &self, source: &str, breakpoints: &ResolvedBreakpoints, ) -> Result<()>

Sets multiple breakpoints for a file (and clears existing ones).

See https://microsoft.github.io/debug-adapter-protocol/specification#Requests_SetBreakpoints

fn top_frame(&self) -> Result<Option<StackFrame>>

Gets the top stack frame, may be None if entered from native.

fn stack_trace( &self, args: StackTraceArguments, ) -> Result<StackTraceResponseBody>

Gets a stacktrace from the current execution state.

See https://microsoft.github.io/debug-adapter-protocol/specification#Requests_StackTrace

fn scopes(&self) -> Result<ScopesInfo>

Gets the variables scope for a frame.

See https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Scopes

fn variables(&self) -> Result<VariablesInfo>

Gets variables for the current scope

See https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Variables

fn inspect_variable(&self, path: VariablePath) -> Result<InspectVariableInfo>

Gets all child variables for the given access path

See https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Variables

fn continue_(&self) -> Result<()>

Resumes execution.

See https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Continue

fn step(&self, kind: StepKind) -> Result<()>

Continues execution until some condition.

See https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Next https://microsoft.github.io/debug-adapter-protocol/specification#Requests_StepIn https://microsoft.github.io/debug-adapter-protocol/specification#Requests_StepOut

fn evaluate(&self, expr: &str) -> Result<EvaluateExprInfo>

Evaluates in expression in the context of the top-most frame.

See https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Evaluate