starlark::debug

Trait DapAdapterEvalHook

pub trait DapAdapterEvalHook:
    Debug
    + Send
    + 'static {
    // Required method
    fn add_dap_hooks(self: Box<Self>, eval: &mut Evaluator<'_, '_, '_>);
}

This is sort of the evaluation side of the DapAdapter. It’s expected that these are on different threads (the starlark evaluation is single-threaded, so certainly the DapAdapter itself doesn’t do interesting things there).

Required Methods

fn add_dap_hooks(self: Box<Self>, eval: &mut Evaluator<'_, '_, '_>)

Hooks the evaluator for this DapAdapter.