starlark

Module environment

Types representing Starlark modules (Module and FrozenModule) and global variables (Globals).

Global functions and values are stored in Globals, which are typically built using GlobalsBuilder. User executions store their values in a Module, which have to be converted to a FrozenModule using freeze before they can be load()’d as a dependency.

Structs

FrozenModule

The result of freezing a Module, making it and its contained values immutable.

Globals

The global values available during execution.

GlobalsBuilder

Used to build a Globals value.

GlobalsStatic

Used to create globals.

Methods

Methods of an object.

MethodsBuilder

Used to build a Methods value.

MethodsStatic

Used to create methods for a StarlarkValue.

Module

A container for user values, used during execution.

Enums

LibraryExtension

The extra library definitions available in this Starlark implementation, but not in the standard.