starlark

Module typing

Types required to support the typecheck function.

Structs

Approximation

A typing operation wasn’t able to produce a precise result, so made some kind of approximation.

Interface

Interface representing the types of all bindings in a module.

ParamSpec

Callable parameter specification (e.g. positional only followed by **kwargs).

Ty

A Starlark type.

TyCallable

typing.Callable.

TyFunction

A function.

TyStarlarkValue

Type implementation where typing is handled by the StarlarkValue trait implementation.

TyStruct

Struct type.

TyUser

Type description for arbitrary type.

TyUserFields

Fields of the struct.

TyUserIndex

Types of [] operator.

TyUserParams

Optional parameters to TyUser::new.

TypeMap

Structure containing all the inferred types.

TypingOracleCtx

Oracle reference with utility methods.

Enums

ParamIsRequired

Indication whether parameter is required.

TyBasic

Type that is not a union.

TypeRenderConfig

Configuration for rendering types.

TypingBinOp

Binary operator for typechecker.

TypingUnOp

Unary operator for typechecker.

Traits

AstModuleTypecheck

Typecheck a module.