starlark::values

Struct FrozenValue

pub struct FrozenValue(/* private fields */);

A Value that can never be changed. Can be converted back to a Value with to_value.

A FrozenValue exists on a FrozenHeap, which in turn can be kept alive by a FrozenHeapRef. If the frozen heap gets dropped while a FrozenValue from it still exists, the program will probably segfault, so be careful when working directly with FrozenValues. See the type OwnedFrozenValue for a little bit more safety.

Implementations

impl FrozenValue

pub fn new_none() -> Self

Create a new value representing None in Starlark.

pub fn new_bool(x: bool) -> Self

Create a new boolean in Starlark.

pub fn new_empty_list() -> Self

Create a new empty list.

pub fn new_empty_dict() -> Self

Create a new empty dict.

pub fn is_none(self) -> bool

Is a value a Starlark None.

pub fn unpack_bool(self) -> Option<bool>

Return the bool if the value is a boolean, otherwise None.

pub fn unpack_i32(self) -> Option<i32>

Obtain the underlying integer if it fits in an i32. Note floats are not considered integers, i. e. unpack_i32 for 1.0 will return None.

pub fn to_value<'v>(self) -> Value<'v>

Convert a FrozenValue back to a Value.

pub fn downcast_frozen_ref<T: StarlarkValue<'static>>( self, ) -> Option<FrozenRef<'static, T>>

Downcast to given type.

pub fn downcast_frozen_str(self) -> Option<FrozenRef<'static, str>>

Downcast to string.

pub fn downcast_frozen_starlark_str( self, ) -> Option<FrozenRef<'static, StarlarkStr>>

Note: see docs about [‘Value::unpack_box_str’] about instability