starlark::values

Struct OwnedFrozenRef

pub struct OwnedFrozenRef<T: ?Sized + 'static> { /* private fields */ }

Same as a FrozenRef, but it keeps itself alive by storing a reference to the owning heap.

Usually constructed from an OwnedFrozenValueTyped.

Implementations

impl<T: ?Sized> OwnedFrozenRef<T>

pub unsafe fn new_unchecked( value: &'static T, owner: FrozenHeapRef, ) -> OwnedFrozenRef<T>

Creates a new OwnedFrozenRef pointing at the given value.

Safety

The reference must be kept alive by the owning heap

pub fn as_owned_ref_frozen_ref(&self) -> OwnedRefFrozenRef<'_, T>

Borrow.

pub fn as_ref<'a>(&'a self) -> &'a T

Returns a reference to the underlying value.

pub fn map<F, U: ?Sized>(self, f: F) -> OwnedFrozenRef<U>
where for<'v> F: FnOnce(&'v T) -> &'v U,

Converts self into a new reference that points at something reachable from the previous.

See the caveats on [starlark::values::OwnedFrozenValue::map]

pub fn try_map_result<F, U: ?Sized, E>( self, f: F, ) -> Result<OwnedFrozenRef<U>, E>
where for<'v> F: FnOnce(&'v T) -> Result<&'v U, E>,

Fallible map the reference to another one.

pub fn try_map_option<F, U: ?Sized>(self, f: F) -> Option<OwnedFrozenRef<U>>
where for<'v> F: FnOnce(&'v T) -> Option<&'v U>,

Optionally map the reference to another one.

pub fn owner(&self) -> &FrozenHeapRef

Get a reference to the owning frozen heap