starlark::values

Trait AllocFrozenValue

pub trait AllocFrozenValue: StarlarkTypeRepr {
    // Required method
    fn alloc_frozen_value(self, heap: &FrozenHeap) -> FrozenValue;
}

Trait for things that can be allocated on a FrozenHeap producing a FrozenValue.

Derive

AllocFrozenValue can be derived for enums, like this:

use starlark::values::AllocFrozenValue;
use starlark::values::type_repr::StarlarkTypeRepr;

#[derive(StarlarkTypeRepr, AllocFrozenValue)]
enum AllocIntOrStr {
    Int(i32),
    Str(String),
}

Required Methods

fn alloc_frozen_value(self, heap: &FrozenHeap) -> FrozenValue

Allocate a value in the frozen heap and return a reference to the allocated value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.