starlark::values

Trait AllocValue

pub trait AllocValue<'v>: StarlarkTypeRepr {
    // Required method
    fn alloc_value(self, heap: &'v Heap) -> Value<'v>;
}

Trait for things that can be created on a Heap producing a Value.

Note, this trait does not represent Starlark types. For example, this trait is implemented for char, but there’s no Starlark type for char, this trait is implemented for char to construct Starlark str.

For types that implement crate::values::StarlarkValue a typical implementation will probably call either Heap::alloc_simple or Heap::alloc_complex, e.g.

#[derive(Debug, derive_more::Display, Allocative, NoSerialize, ProvidesStaticType)]
struct MySimpleValue;

#[starlark_value(type = "MySimpleValue", UnpackValue, StarlarkTypeRepr)]
impl<'v> StarlarkValue<'v> for MySimpleValue {}

impl<'v> AllocValue<'v> for MySimpleValue {
    fn alloc_value(self, heap: &'v Heap) -> Value<'v> {
        heap.alloc_simple(self)
    }
}

Derive

AllocValue can be derived for enums, like this:

use starlark::values::AllocValue;
use starlark::values::type_repr::StarlarkTypeRepr;

#[derive(StarlarkTypeRepr, AllocValue)]
enum AllocIntOrStr {
    Int(i32),
    Str(String),
}

Required Methods

fn alloc_value(self, heap: &'v Heap) -> Value<'v>

Allocate the value on a heap and return a reference to the allocated value.

Note, for certain values (e.g. empty strings) no allocation is actually performed, and a reference to the statically allocated object is returned.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.