starlark::values

Trait UnpackValue

pub trait UnpackValue<'v>: Sized + StarlarkTypeRepr {
    type Error: UnpackValueError;

    // Required method
    fn unpack_value_impl(value: Value<'v>) -> Result<Option<Self>, Self::Error>;

    // Provided methods
    fn unpack_value(value: Value<'v>) -> Result<Option<Self>, Error> { ... }
    fn unpack_value_opt(value: Value<'v>) -> Option<Self>
       where Self::Error: UnpackValueErrorInfallible { ... }
    fn unpack_value_err(value: Value<'v>) -> Result<Self> { ... }
    fn unpack_param(value: Value<'v>) -> Result<Self> { ... }
    fn unpack_named_param(value: Value<'v>, param_name: &str) -> Result<Self> { ... }
}

How to convert a Value to a Rust type. Required for all arguments in a #[starlark_module] definition.

Note for simple references it often can be implemented with #[starlark_value(UnpackValue)], for example:

#[derive(
    Debug,
    derive_more::Display,
    Allocative,
    NoSerialize,
    ProvidesStaticType
)]
struct MySimpleValue;

#[starlark_value(type = "MySimpleValue", UnpackValue, StarlarkTypeRepr)]
impl<'v> StarlarkValue<'v> for MySimpleValue {}

Whereas for types that aren’t also StarlarkValue you can define:

struct BoolOrInt(i32);

impl StarlarkTypeRepr for BoolOrInt {
    type Canonical = <Either<bool, i32> as StarlarkTypeRepr>::Canonical;

    fn starlark_type_repr() -> Ty {
        Either::<bool, i32>::starlark_type_repr()
    }
}

impl<'v> UnpackValue<'v> for BoolOrInt {
    type Error = starlark::Error;

    fn unpack_value_impl(value: Value<'v>) -> starlark::Result<Option<Self>> {
        if let Some(x) = value.unpack_bool() {
            Ok(Some(BoolOrInt(x as i32)))
        } else {
            let Some(x) = i32::unpack_value(value)? else {
                return Ok(None);
            };
            Ok(Some(BoolOrInt(x)))
        }
    }
}

Required Associated Types

type Error: UnpackValueError

Error returned when type matches, but conversion fails.

Typically starlark::Error, anyhow::Error, or Infallible.

Required Methods

fn unpack_value_impl(value: Value<'v>) -> Result<Option<Self>, Self::Error>

Given a Value, try and unpack it into the given type, which may involve some element of conversion.

Return None if the value is not of expected type (as described by StarlarkTypeRepr, and return Err if the value is of expected type, but conversion cannot be performed. For example, when unpacking an integer to String, return None, and when unpacking a large integer to i32, return Err.

This function is needs to be implemented, but usually not meant to be called directly. Consider using unpack_value, unpack_value_err, unpack_value_opt instead.

Provided Methods

fn unpack_value(value: Value<'v>) -> Result<Option<Self>, Error>

Given a Value, try and unpack it into the given type, which may involve some element of conversion.

Return None if the value is not of expected type (as described by StarlarkTypeRepr, and return Err if the value is of expected type, but conversion cannot be performed. For example, when unpacking an integer to String, return None, and when unpacking a large integer to i32, return Err.

fn unpack_value_opt(value: Value<'v>) -> Option<Self>
where Self::Error: UnpackValueErrorInfallible,

Unpack a value if unpacking is infallible.

fn unpack_value_err(value: Value<'v>) -> Result<Self>

Unpack a value, but return error instead of None if unpacking fails.

fn unpack_param(value: Value<'v>) -> Result<Self>

Unpack value, but instead of None return error about incorrect argument type.

fn unpack_named_param(value: Value<'v>, param_name: &str) -> Result<Self>

Unpack value, but instead of None return error about incorrect named argument type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.