starlark::values::type_repr

Trait StarlarkTypeRepr

pub trait StarlarkTypeRepr {
    type Canonical: StarlarkTypeRepr;

    // Required method
    fn starlark_type_repr() -> Ty;
}

Provides a starlark type representation, even if StarlarkValue is not implemented.

Derive

There is #[derive(StarlarkTypeRepr)] for enums, for example:

use starlark::values::type_repr::StarlarkTypeRepr;

#[derive(StarlarkTypeRepr)]
enum IntOrString {
    Int(i32),
    String(String),
}

It emits type int | str.

This derive is useful in combination with derive of UnpackValue.

Required Associated Types

type Canonical: StarlarkTypeRepr

Different Rust type representing the same Starlark Type.

For example, bool and StarlarkBool Rust types represent the same Starlark type bool.

Formal requirement: Self::starlark_type_repr() == Self::Canonical::starlark_type_repr().

If unsure, it is safe to put = Self here. When associated_type_defaults is stabilized, this will be the default.

Required Methods

fn starlark_type_repr() -> Ty

The representation of a type that a user would use verbatim in starlark type annotations

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.