starlark::values

Module typing

Typechecker-related types.

Structs

FrozenStarlarkCallable

Marker for a callable value.

StarlarkCallable

Marker for a callable value. Can be used in function signatures for better documentation and type checking.

StarlarkCallableChecked

More strict version of StarlarkCallable.

StarlarkCallableParamAny

Indicates that a callable accepts any number of positional and keyword arguments.

StarlarkIter

StarlarkTypeRepr for iterable types.

TypeCompiled

Wrapper for a Value that acts like a runtime type matcher.

TypeInstanceId

Globally unique identifier for a type, like record type or enum type.

TypeMatcherFactory

Boxed TypeMatcher.

TypeType

Represent a type of type. (For example, an expression int is valid for this type.)

Enums

StarlarkNever

Never type, can be used as native function return type.

Traits

StarlarkCallableParamSpec

Type parameter for StarlarkCallable or FrozenStarlarkCallable describing the expected parameters of the callable.

TypeMatcher

Runtime type matcher. E.g. when isinstance(1, int) is called, implementation of TypeMatcher for int is used.