starlark::values::typing

Struct StarlarkCallableChecked

pub struct StarlarkCallableChecked<'v, P: StarlarkCallableParamSpec, R: StarlarkTypeRepr>(pub Value<'v>, _);

More strict version of StarlarkCallable.

This checks not only that the value is callable, but also that it is a callable with the correct signature.

The implementation uses starlark-rust typechecker with all its limitations. For example, if there are optional parameters in both value-def and this signature, signature matching is ignored at the time of writing.

Unpacking with this type is expensive: usually it is OK to use it for code executed once at top-level scope (like rule()), but not for code executed many times (like partial()).

Tuple Fields
0: Value<'v>
Implementations

impl<'v, P: StarlarkCallableParamSpec, R: StarlarkTypeRepr> StarlarkCallableChecked<'v, P, R>

pub fn to_unchecked(self) -> StarlarkCallable<'v, P, R>

Convert to StarlarkCallable.